Churn Analysis
The Telco customer churn data set was used to for this predictive
analytics exercise. The data has 7,043 observations and 21 variables.
Three prediction models were explored: Logit, SVM, and GBM. A t-test
evaluation was conducted which showed that none of the three models had
a superior accuracy over the other. However, the Logit model appeared to
have the highest computed Kappa statistic.